move types into Component files instead of individual files
