--- application/react/_types/forms/address.ts
@@ -1,8 +1,0 @@
-export type AddressFormAttributes = {
-  name: string,
-  street: string,
-  city: string,
-  state: string,
-  zipCode: string,
-  unitOrSuite?: string
-}
