--- application/react/_types/forms/adhesiveCategory.ts
@@ -1,3 +1,0 @@
-export type AdhesiveCategoryFormAttributes = {
-  name: string; 
-}
