--- application/react/_types/forms/creditTerm.ts
@@ -1,3 +1,0 @@
-export type CreditTermFormAttributes = {
-  description: string
-}
