--- application/react/_types/forms/deliveryMethod.ts
@@ -1,3 +1,0 @@
-export type DeliveryMethodFormAttributes = {
-  name: string
-}
