--- application/react/_types/forms/linerType.ts
@@ -1,3 +1,0 @@
-export type LinerTypeForm = {
-  name: String
-}
