--- application/react/_types/forms/materialLengthAdjustment.ts
@@ -1,7 +1,0 @@
-import { MongooseId } from "../typeAliases"
-
-export type MaterialLengthAdjustmentFormFormAttributes = {
-  material: MongooseId,
-  length: number,
-  notes?: string
-}
