--- application/react/_types/forms/shippingLocation.ts
@@ -1,7 +1,0 @@
-import { MongooseId } from "../typeAliases";
-import { AddressFormAttributes } from "./address";
-
-export type ShippingLocationFormAttributes = AddressFormAttributes & {
-  freightAccountNumber?: string,
-  deliveryMethod?: MongooseId,
-}
